requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
decorator==5.1.1
Django==4.0.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.2
future==0.18.2
idna==3.3
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
pytz==2021.3
ratelim==0.1.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
yandex-geocoder==2.0.0
